% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_helpers.R
\name{ggsave_mpinl_poster_portrait}
\alias{ggsave_mpinl_poster_portrait}
\alias{ggsave_mpinl_poster_landscape}
\title{Saving helper functions}
\usage{
plot + theme_mpinl_poster()
ggsave_mpinl_poster_portrait("posterfigure_1.pdf") # The default height is
ggsave_mpinl_poster_portrait("posterfigure_2.pdf",height = 200, units = "mm")
ggsave_mpinl_poster_landscape("outputfilename_landscape.pdf")

ggsave_mpinl_poster_landscape(filename, plot = ggplot2::last_plot(),
  device = NULL, path = NULL, scale = 1, width = NA, height = NA,
  units = c("in", "cm", "mm"), dpi = 500, limitsize = TRUE, ...)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{device}{Device to use. Can be either be a device function
(e.g. \code{\link[=png]{png()}}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{path}{Path to save plot to (combined with filename).}

\item{scale}{Multiplicative scaling factor.}

\item{width}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{height}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{units}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{limitsize}{When \code{TRUE} (the default), \code{ggsave} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}

\item{...}{Other arguments passed on to the graphics device function,
as specified by \code{device}.}
}
\description{
These functions are wrappers around the ggplot2 convenience function \code{\link[ggplot2]{ggsave}}, for saving a plot.
They default to saving the last plot that you displayed.
The width is in each case predefined to have the correct width for
the MPI poster templates. Do adjust the height though!
There are versions for portrait and landscape poster templates, since these have different column widths.
}
\examples{
mtcars2 <- within(mtcars, {
vs <- factor(vs, labels = c("V-shaped", "Straight"))
am <- factor(am, labels = c("Automatic", "Manual"))
cyl  <- factor(cyl)
gear <- factor(gear)
})

library(ggplot2)
library(mpinltheme)

p1 <- ggplot(mtcars2) +
  geom_point(aes(x = wt, y = mpg, colour = gear)) +
    labs(title = "Fuel economy declines as weight increases",
         subtitle = "(1973-74)",
         x = "Weight (1000 lbs)",
         y = "Fuel economy (mpg)",
         colour = "Gears")

p1 + scale_colour_discrete_mpinl() + theme_mpinl_poster()
ggsave_mpinl_poster_portrait("outputfilename.pdf")


}
