% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{mpinl_colours}
\alias{mpinl_colours}
\alias{mpinl_colors}
\title{Colours for the MPI for Psycholinguistics ggplot2 theme
This object contains a named character vector of
colours consistent with the the MPI for Psycholinguistics
corporate identity.}
\format{An object of class \code{character} of length 24.}
\usage{
mpinl_colours

mpinl_colors
}
\description{
Colours for the MPI for Psycholinguistics ggplot2 theme
This object contains a named character vector of
colours consistent with the the MPI for Psycholinguistics
corporate identity.
}
\keyword{datasets}
