% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_mpinl_slide}
\alias{theme_mpinl_slide}
\title{MPI for Psycholinguistics ggplot2 theme, version for slides}
\usage{
plot + scale_colour_discrete_mpinl() + theme_mpinl_slide()
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Theme to produce plots that play nicely with the slide templates for the MPI for Psycholinguistics corporate identity.

There is another version available, \code{\link{theme_mpinl_general}}, with text sizes for other purposes.

This theme can be appended to a ggplot2 plot object.
The theme is intended to be used in combination with \code{\link{scale_colour_discrete_mpi_nl}} and \code{\link{scale_fill_discrete_mpi_nl}}
to use the recommended colour scheme. For continuous colour scales, we recommend \code{\link[viridis]{scale_color_viridis}}.

It shouldn't be necessary to tweak the further. You might wish to use additional + theme() calls to
further customise the appearance of your plots, for instance to control the placement of the legend.
}
\examples{
mtcars2 <- within(mtcars, {
vs <- factor(vs, labels = c("V-shaped", "Straight"))
am <- factor(am, labels = c("Automatic", "Manual"))
cyl  <- factor(cyl)
gear <- factor(gear)
})

library(ggplot2)

p1 <- ggplot(mtcars2) +
  geom_point(aes(x = wt, y = mpg, colour = gear)) +
    labs(title = "Fuel economy declines as weight increases",
         subtitle = "(1973-74)",
         x = "Weight (1000 lbs)",
         y = "Fuel economy (mpg)",
         colour = "Gears")

p1 + scale_colour_discrete_mpinl() + theme_mpinl_slide()

}
